Fix tremendous bug in sbt-release
After hours chasing this bug, this commit finally fixes the publishing
of sbt subprojects independently. For some reason, when executing
`runClean`, we create a new state that does not have the settings of the
previous state, and then the ivy file is not correctly delivered. When
it comes to releasing the package, sbt bails out because that ivy file
is not in the right place.